WIP: Upgrade to Apollo Server v2
