+++ api/models/usersChannels.js
@@ -131,7 +131,7 @@ const removeMemberInChannel = (channelId: string, userId: string): Promise<?DBCh
             context: { channelId }
           })
         }
-        
+
         return db.table('channels').get(join.channelId).run();
       } else {
         return null;
@@ -340,7 +340,7 @@ const approvePendingUsersInChannel = (channelId: string): Promise<DBUsersChannel
 // as a member
 // prettier-ignore
 const approveBlockedUserInChannel = (channelId: string, userId: string): Promise<DBUsersChannels> => {
-  
+
   trackQueue.add({
     userId,
     event: events.USER_WAS_UNBLOCKED_IN_CHANNEL,
@@ -479,7 +479,7 @@ const toggleUserChannelNotifications = async (userId: string, channelId: string,
     .getAll(channelId, { index: 'channelId' })
     .filter({ userId })
     .run();
-  
+
   // permissions exist, this user is trying to toggle notifications for a channel where they
   // are already a member
   if (permissions && permissions.length > 0) {
