+++ api/package.json
@@ -3,6 +3,7 @@
     "algoliasearch": "^3.30.0",
     "apollo-engine": "1.x",
     "apollo-local-query": "^0.3.0",
+    "apollo-server-express": "^2.0.4",
     "apollo-upload-client": "^8.1.0",
     "apollo-upload-server": "^5.0.0",
     "aws-sdk": "2.200.0",
@@ -46,7 +47,6 @@
     "graphql-date": "^1.0.3",
     "graphql-depth-limit": "^1.1.0",
     "graphql-log": "0.1.2",
-    "graphql-server-express": "1.3.0",
     "graphql-subscriptions": "0.4.x",
     "graphql-tools": "3.0.0",
     "helmet": "^3.13.0",
