+++ api/routes/api/index.js
@@ -3,11 +3,11 @@ import { Router } from 'express';
 
 const apiRouter = Router();
 
-import graphiql from './graphiql';
-// Only allow GraphiQL in development
-if (process.env.NODE_ENV === 'development') {
-  apiRouter.use('/graphiql', graphiql);
-}
+// import graphiql from './graphiql';
+// // Only allow GraphiQL in development
+// if (process.env.NODE_ENV === 'development') {
+//   apiRouter.use('/graphiql', graphiql);
+// }
 
 import slackImporter from './slackImporter';
 apiRouter.use('/slack', slackImporter);
@@ -21,7 +21,7 @@ apiRouter.use('/email', email);
 import userExportRouter from './export-user-data';
 apiRouter.use('/user.json', userExportRouter);
 
-import graphql from './graphql';
-apiRouter.use('/', graphql);
+// import graphql from './graphql';
+// apiRouter.use('/', graphql);
 
 export default apiRouter;
