+++ package.json
@@ -56,6 +56,7 @@
     "apollo-link-http": "^1.5.4",
     "apollo-link-retry": "^2.2.4",
     "apollo-link-ws": "^1.0.8",
+    "apollo-server-express": "^2.0.4",
     "apollo-upload-client": "^8.1.0",
     "apollo-upload-server": "^5.0.0",
     "apollo-utilities": "^1.0.16",
@@ -102,7 +103,6 @@
     "graphql-date": "^1.0.3",
     "graphql-depth-limit": "^1.1.0",
     "graphql-log": "0.1.3",
-    "graphql-server-express": "1.3.6",
     "graphql-subscriptions": "0.5.8",
     "graphql-tag": "^2.9.2",
     "graphql-tools": "^3.0.5",
