+++ shared/middlewares/cors.js
@@ -1,7 +1,7 @@
 // @flow
 import cors from 'cors';
 
-export default cors({
+export const corsOptions = {
   origin:
     process.env.NODE_ENV === 'production' && !process.env.FORCE_DEV
       ? [
@@ -13,4 +13,6 @@ export default cors({
         ].filter(Boolean)
       : [/localhost/],
   credentials: true,
-});
+};
+
+export default cors(corsOptions);
